Refactor: Moved useGraphMarketsFromQuestion inside hooks/Graph folder
