+++ app/src/hooks/Graph/useGraphMarketsFromQuestion.tsx
@@ -2,8 +2,8 @@ import { useQuery } from '@apollo/react-hooks'
 import gql from 'graphql-tag'
 import { useEffect, useState } from 'react'
 
-import { isObjectEqual } from '../util/tools'
-import { Status } from '../util/types'
+import { isObjectEqual } from '../../util/tools'
+import { Status } from '../../util/types'
 
 const query = gql`
   query GetQuestion($id: ID!) {
